Move token arrayb in a new file
